My somewhat embarrassing attempt at miming DCT
Given the book's input, my output is near-identical to the book's
output. The first value in the DCT seems extraordinarily large, and the
final value in the first row is negative when it should not be. All
other numbers being accurate, I figured it was time to upload to GitHub
before I start testing changes that don't work.