Started Decompression.py, tweaked Decode_Width()
Minor tweaks to Compression.py as well, such as Test_Run_Width()
slightly to allow the code to tell the user if it thinks it decodes
properly, rather than leaving it up to the eye.

Decode_Width() no longer returns a string look-alike to the original
list; now actually returns a list.