Fixed Run_Width() Bug! Haha!
Aha! I KNEW the order of the while loops in Run_Width would cause
issues! Intuitively, I was using them akin to if statements, but  I felt
hybrid elif-while loops wouldn't make any sense/would be redundant, I
knew that my intuitive sense was going to be jaded, unable to see
potential problems... I knew it! I knew I'd be blind to the problems
they might cause in the future and I was right. My intuition knew it may
contain an error, and was totally right. Wow. Human intuition is crazy.
Sorry, I'm a little tired (and excited that bug didn't take much longer
to fix.)

Debug.py:
Aesthetic changes
Decompression.py:
Removed an outdated comment
Compression.py:
Introduced seemingly-redundant if/elif-while hybrids.
(The problem was that by de-incrementing the values in the first while
loop, it was accidentally possible to enter the second, meaning an
additional character would be inserted that caused the decoder to read
the line incorrectly.)
(The solution was NOT to swap them, so that the condition to pass into
the loop is met too late to enter, because the reverse- the (now) first
loop accidentally entering the (now) second- would happen instead.)
(The solution was to make them mutually exclusive, mostly because that's
a fun word.)