Added Resaver.py
So images that go in do not look the same when coming out. My goal is to
figure out why.
I'm going to slowly add Compression/Decompression functions to
Resaver.py until I get to the root of the problem.
I will never upload a version of Resaver.py that does NOT re-save input
images precisely (enough) as they were found unless explicitly stated
otherwise. So, currently, it takes in an image, and the output is
identical. Again, will be slowly adding Compression/Decompression
functions until the images start to change so I can isolate the problem.