Changed my mind about the shift in code structure
Not that I intend to forsake it completely; I want to focus on the
PyCUDA and threading FIRST before trying to clean up everything, simply
because changing code structure to a "cleaner" also means the functions
themselves have to be a little messier in ways that are simply
unnecssary. Originally, making these functions "messy" was part of the
transition to threading (and to PyCUDA) but I realize now that's just
frivolous; I already understand how I would make those changes, so I
should just go for it rather than waste time.